refactored listing and fixed UI
