+++ Frontend/src/app/listings/page.js
@@ -1,156 +1,16 @@
 "use client";
 
+import FilterSidebar from "@/components/listings/filterSidebar";
+import haversineDistanceMiles from "@/components/listings/haversineDistance";
 import ListingCard from "@/components/listings/listingCard";
 import SearchBar from "@/components/listings/searchBar";
 import ListingDetailModal from "@/components/listings/listingDetailModal";
 import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
-import { LucideIcon, Star } from "lucide-react";
 import MapDropLocation from "@/components/listings/mapDropLocation";
 import BookingModal from "@/components/listings/bookingModal";
 import { CreateListingModal } from "@/components/listings/createListingModal";
 import { fetchListings } from "@/api/fetchListings";
-
-// this may eventually be complex enough to be pulled into its own component file
-function LeftBar({
-  showBikes,
-  setShowBikes,
-  showScooters,
-  setShowScooters,
-  price,
-  setPrice,
-  rating,
-  setRating,
-  distance,
-  setDistance,
-}) {
-  const handleBikeChange = () => {
-    setShowBikes(!showBikes);
-  };
-  const handleScooterChange = () => {
-    setShowScooters(!showScooters);
-  };
-  const handleStars = (value) => {
-    setRating(value);
-  };
-
-  return (
-    <div className="w-64 h-[500px] bg-lighterGray rounded-r-3xl overflow-hidden mt-8 border border-lightGray">
-      <div className="w-full bg-waxwingGreen text-white text-2xl font-bold flex items-center justify-center gap-2 h-14.5 rounded-tr-3xl">
-        <img
-          src="filter-svgrepo-com.svg"
-          alt="Filter Icon"
-          className="h-[1em] w-[1em] align-middle"
-        />
-        <span className="align-middle">Filters</span>
-      </div>
-      <div className="space-y-2 p-4 pb-8">
-        <div className="w-full h-6 text-center justify-center text-lg font-bold">
-          Ride Type
-        </div>
-        <label className="flex items-center justify-start space-x-6 w-full pl-6">
-          <input
-            type="checkbox"
-            checked={showBikes}
-            onChange={handleBikeChange}
-            className="accent-[var(--color-waxwingGreen)] w-6 h-6"
-          />
-          <span className="text-lg flex-1 text-left">Bikes</span>
-        </label>
-        <label className="flex items-center justify-start space-x-6 w-full pl-6">
-          <input
-            type="checkbox"
-            checked={showScooters}
-            onChange={handleScooterChange}
-            className="accent-[var(--color-waxwingGreen)] w-6 h-6"
-          />
-          <span className="text-lg flex-1 text-left">Scooters</span>
-        </label>
-      </div>
-
-      <div className="space-y-2 pb-8 w-full">
-        <div className="w-full h-6 text-center justify-center text-lg font-bold">
-          Hourly Price
-        </div>
-        <div className="flex flex-col items-center w-full space-y-2">
-          <input
-            type="range"
-            min="0"
-            max="25"
-            value={price}
-            onChange={(e) => setPrice(Number(e.target.value))}
-            className="w-3/4 h-2 accent-[var(--color-waxwingGreen)] rounded-lg"
-          />
-          <div className="flex justify-mid text-sm font-semibold text-gray-700">
-            <span>${price}</span>
-          </div>
-        </div>
-      </div>
-
-      <div className="space-y-2 pb-8 w-full">
-        <div className="w-full h-6 text-center justify-center text-lg font-bold">
-          Distance
-        </div>
-        <div className="flex flex-col items-center w-full space-y-2">
-          <input
-            type="range"
-            min="0"
-            max="100"
-            step="0.1"
-            value={distance}
-            onChange={(e) => setDistance(Number(e.target.value))}
-            className="w-3/4 h-2 accent-[var(--color-waxwingGreen)] rounded-lg"
-          />
-          <div className="flex justify-center text-sm font-semibold text-gray-700">
-            <span>{distance} Miles</span>
-          </div>
-        </div>
-      </div>
-
-      <div className="space-y-2 w-full flex flex-col items-center">
-        <div className="w-full h-6 text-center justify-center text-lg font-bold">
-          Rating
-        </div>
-        <div className="flex space-x-1">
-          {[1, 2, 3, 4, 5].map((star) => (
-            <Star
-              key={star}
-              fill={rating >= star ? "var(--color-waxwingGreen)" : "#f2f2f2"}
-              strokeWidth={1}
-              // stroke="var(--color-waxwingGreen)"
-              className={`w-8 h-8 cursor-pointer ${
-                rating >= star
-                  ? "text-[var(--color-waxwingGreen)]"
-                  : "text-gray-300"
-              }`}
-              onClick={() => handleStars(star)}
-            />
-          ))}
-        </div>
-        <span className="flex justify-center text-sm font-semibold text-gray-700">
-          {rating > 0
-            ? `${rating} star${rating > 1 ? "s" : ""}+`
-            : "Any rating"}
-        </span>
-      </div>
-    </div>
-  );
-}
-
-function haversineDistanceMiles(lat1, lon1, lat2, lon2) {
-  const R = 3958.8; // Earth radius in miles
-  const dLat = ((lat2 - lat1) * Math.PI) / 180;
-  const dLon = ((lon2 - lon1) * Math.PI) / 180;
-
-  const a =
-    Math.sin(dLat / 2) ** 2 +
-    Math.cos((lat1 * Math.PI) / 180) *
-      Math.cos((lat2 * Math.PI) / 180) *
-      Math.sin(dLon / 2) ** 2;
-
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-  return R * c; // distance in miles
-}
 
 export default function ListingsPage() {
   const [price, setPrice] = useState(25);
@@ -282,8 +142,7 @@ export default function ListingsPage() {
           handler={() => fetchListings(setListings)}
         />
       )}
-      {/* Filter sidebar */}
-      <LeftBar
+      <FilterSidebar
         showBikes={showBikes}
         setShowBikes={setShowBikes}
         showScooters={showScooters}
@@ -295,7 +154,6 @@ export default function ListingsPage() {
         distance={maxDistance}
         setDistance={setMaxDistance}
       />
-
       <div className="flex-1 flex flex-col overflow-hidden">
         {/* Search bar, fixed at top */}
         <div className="flex-shrink-0 p-6 bg-white border-b border-gray-200 flex justify-center gap-x-4">
@@ -316,7 +174,7 @@ export default function ListingsPage() {
         */}
         <main className="flex-1 overflow-y-auto">
           <div className="p-6 flex justify-center">
-            <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-6justify-items-center">
+            <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-16 justify-items-center">
               {/* Render listing cards here based on fetched bikes,
               need to figure out proper fields to pass in based on backend data
               */}
